requirements.txt
pandas==2.3.1
openpyxl==3.1.5
requests==2.32.4
python-dotenv==1.0.1
scikit-learn==1.7.1
